Added json formatter

+++ include/spdlog/common.h
@@ -338,7 +338,7 @@ struct file_event_handlers {
     std::function<void(const filename_t &filename)> after_close;
 };
 
-using FieldValue = std::variant<spdlog::string_view_t, int, long>;  // TODO
+using FieldValue = std::variant<spdlog::string_view_t, int, double>;  // TODO
 struct Field {
     spdlog::string_view_t name;
     FieldValue value;
